Rename the Quartz.Plugin.* namespaces to Quartz.Plugins.*
The packages are Quartz.Plugins and Quartz.Plugins.TimeZoneConverter; only the
namespaces were singular. History, Interrupt, Json, Management, Xml and
TimeZoneConverter all move, so namespace, assembly and package agree.

The quartz.plugin.<name>.* configuration key prefix is unchanged — it is a
property key, not a namespace. The type it names goes through
SimpleTypeLoadHelper, which gains a Quartz.Plugin. -> Quartz.Plugins. entry, so
the old spelling still resolves, with a warning.
