+++ src/Quartz.Examples/10_RunningJobsByPlugInXmlConfiguration/RunningJobsByPlugInXmlConfigurationExample.cs
@@ -38,8 +38,8 @@ public class RunningJobsByPlugInXmlConfigurationExample : IExample
         // and makes it watch for changes every two minutes (120 seconds)
         NameValueCollection properties = new NameValueCollection
         {
-            ["quartz.plugin.triggHistory.type"] = "Quartz.Plugin.History.LoggingJobHistoryPlugin, Quartz.Plugins",
-            ["quartz.plugin.jobInitializer.type"] = "Quartz.Plugin.Xml.XMLSchedulingDataProcessorPlugin, Quartz.Plugins",
+            ["quartz.plugin.triggHistory.type"] = "Quartz.Plugins.History.LoggingJobHistoryPlugin, Quartz.Plugins",
+            ["quartz.plugin.jobInitializer.type"] = "Quartz.Plugins.Xml.XMLSchedulingDataProcessorPlugin, Quartz.Plugins",
             ["quartz.plugin.jobInitializer.fileNames"] = "quartz_jobs.xml",
             ["quartz.plugin.jobInitializer.failOnFileNotFound"] = "true",
             ["quartz.plugin.jobInitializer.scanInterval"] = "120"
