+++ src/Quartz.Server/appsettings.json
@@ -13,7 +13,7 @@
     },
     "Plugin": {
       "Xml": {
-        "Type": "Quartz.Plugin.Xml.XMLSchedulingDataProcessorPlugin, Quartz.Plugins",
+        "Type": "Quartz.Plugins.Xml.XMLSchedulingDataProcessorPlugin, Quartz.Plugins",
         "FileNames": "~/quartz_jobs.xml"
       }
     }
