+++ src/Quartz.Tests.Unit/Plugins/History/RecordingLoggerProvider.cs
