+++ src/Quartz/Impl/SimpleTypeLoadHelper.cs
@@ -44,6 +44,7 @@ internal sealed class SimpleTypeLoadHelper : ITypeLoadHelper
         ("Quartz.Spi.", "Quartz.Extensibility."),
         ("Quartz.Simpl.", "Quartz.Impl."),
         ("Quartz.Job.", "Quartz.Jobs."),
+        ("Quartz.Plugin.", "Quartz.Plugins."),
     ];
 
     /// <summary>
